moved color generation into utils
