+++ lib/view/home_page.dart
@@ -4,6 +4,7 @@ import 'package:wh_covid19/hard_data.dart';
 import 'package:wh_covid19/style.dart';
 import 'package:wh_covid19/view/info_view.dart';
 import 'package:wh_covid19/widget/card_container.dart';
+import 'package:wh_covid19/utils/color.dart';
 
 class HomePage extends StatefulWidget {
   @override
@@ -57,34 +58,6 @@ class _HomePageState extends State<HomePage> {
     );
   }
 
-  Color generateIconColor(
-      Color fromColor, Color toColor, double deltaPercentage) {
-    var fromRed = fromColor.red,
-        fromGreen = fromColor.green,
-        fromBlue = fromColor.blue;
-    var toRed = toColor.red, toGreen = toColor.green, toBlue = toColor.blue;
-
-    var deltaRed = (fromRed - toRed).abs(),
-        deltaGreen = (fromGreen - toGreen).abs(),
-        deltaBlue = (fromBlue - toBlue).abs();
-
-    print('$deltaPercentage $fromColor $toColor');
-    print('$deltaRed $deltaGreen $deltaBlue');
-
-    print(toRed + deltaRed * deltaPercentage);
-    var newRed = fromRed < toRed
-            ? fromRed + deltaRed * deltaPercentage
-            : fromRed - deltaRed * deltaPercentage,
-        newGreen = fromGreen < toGreen
-            ? fromGreen + deltaGreen * deltaPercentage
-            : fromGreen - deltaGreen * deltaPercentage,
-        newBlue = fromBlue < toGreen
-            ? fromBlue + deltaBlue * deltaPercentage
-            : fromBlue - deltaBlue * deltaPercentage;
-
-    return Color.fromRGBO(newRed.toInt(), newGreen.toInt(), newBlue.toInt(), 1);
-  }
-
   Widget _renderBody(BuildContext context) {
     var appBarHeight = 70, logoHeight = 24;
 
@@ -123,8 +96,10 @@ class _HomePageState extends State<HomePage> {
                     mainLogo,
                     IconButton(
                       icon: Icon(Icons.info_outline,
-                          color: generateIconColor(AppColors.homeAppBarIcon,
-                              AppColors.appBarIcon, percentage)),
+                          color: ColorUtils.generateIconColor(
+                              AppColors.homeAppBarIcon,
+                              AppColors.appBarIcon,
+                              percentage)),
                       onPressed: () => InfoView.navigateTo(context),
                     )
                   ],
